Set strict=True in all uses of zip(). Refactored folder name types -> owntypes to adhere to A005 error in ruff updated to ver 0.9.0
