+++ openseries/owntypes.py
