+++ pyproject.toml
@@ -66,7 +66,7 @@ mypy = "1.14.1"
 pandas-stubs = ">=2.1.2,<3.0.0"
 pre-commit = ">=3.7.1,<6.0.0"
 pytest = ">=8.2.2,<9.0.0"
-ruff = "0.8.6"
+ruff = "0.9.0"
 types-openpyxl = ">=3.1.2,<4.0.0"
 types-python-dateutil = ">=2.8.2,<3.0.0"
 types-requests = ">=2.20.0,<3.0.0"
